requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
PyJWT==2.8.0


# Export functionality
fpdf2==2.7.9
python-docx==1.1.0
